AnyGivenPsycho Work Day 2
During this project time I focused mainly on the aspect of using a
magic word. I was able to successfully create a magic word that works
all i have to do is incorporate it into the scenery and actual story. I
created more rooms this time around and am slowly forming a more
cohesive story/map for my game. I was able to help people out when it
came to magic words and renaming NPCs.