AnyGivenPsyco Work Day 3
This is the third day check point. This commit is extremely weak since
most of the class I was working on finding ways to create the desired
puzzle. A secret door combined with a magic word. I will commit later
tonight with the finished puzzle and room designs please don’t grade
this or judge me.