Any Given Psycho Beta Version
This version of the game is my first Beta version. It has all three
puzzles completed, all the items and rooms designed and described. It
also has a working NPC and one ending. This version is complete
however, I intend to ad quite a few modifications after I get the
responses from the beta testers. I want to probably add a second or
better ending to the story and revise or add more puzzles to the mix.
Hopefully things go well. To find the easiest way to beat the game,
scroll to the very bottom of the source code, there in green are the
steps in order to completing the game in its simplest form.