Any Given Psycho Final Work Day
During this day I worked mainly on addressing the issues that the Beta
testers brought up. I intend to make changes and fix up the rest of the
game by 11:59 tonight. This commit mainly deals with some minor
changes, the rest of the class was used to better understand what I
have to turn in.